+++ source/Coordinator.cpp
@@ -306,6 +306,7 @@ void Coordinator::runBenchmarks()
         BenchPhaseConfig { BenchPhase_GET_S3_BUCKET_MD, progArgs.getRunS3GetBucketMetadata() },
 		BenchPhaseConfig { BenchPhase_CREATEFILES, progArgs.getRunCreateFilesPhase() },
         BenchPhaseConfig { BenchPhase_ABORT_MPU, progArgs.getRunAbortMPUPhase() },
+        BenchPhaseConfig { BenchPhase_COPYOBJECT, progArgs.getRunS3CopyObjPhase() },
         BenchPhaseConfig { BenchPhase_PUTOBJACL, progArgs.getRunS3AclPut() },
         BenchPhaseConfig { BenchPhase_PUT_S3_OBJECT_MD, progArgs.getRunS3PutObjectMetadata() },
 		BenchPhaseConfig { BenchPhase_STATFILES, progArgs.getRunStatFilesPhase() },